Rename roundtrip tests
